Set maximum delay for stop bit.
In my environment, 433MHz is quite noisy and there are
plenty of spikes. This triggers in state -1, with a big duration (~65000).
Thus, the min1Period gets a big value, and all of the received pulses get skipped
because "too short". The state machine stays forever in state 0, with all pulses skipped.
Setting a maximum duration value prevents this livelock.